requirements.txt
# PyTorch & Vision
torch>=1.12.1
torchvision>=0.13.1
timm>=0.9.2

# Core libraries
opencv-python>=4.5.5
numpy>=1.18
PyYAML>=5.3
tqdm>=4.62
Pillow>=8.0.0

# COCO support
pycocotools>=2.0.2

# Config & tracking
omegaconf>=2.0.0
clearml>=1.10.0

# Training wrapper
pytorch-lightning>=2.0.0

# ONNX export & conversion
onnx>=1.17.0
onnx2tf==1.22.3

# TensorFlow + Keras
tensorflow==2.16.1
tf-keras==2.16.0
